+++ packages/smarthr-ui/src/components/Dialog/Dialog.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import React, { ComponentProps } from 'react'
 
 import { DialogContentInner } from './DialogContentInner'